requirements.txt
absl-py==1.3.0
altair==4.2.0
apturl==0.5.2
arrow==0.15.5
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
Automat==0.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
flatbuffers==22.10.26
fonttools==4.39.3
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.27.0
imgurpython==1.1.7
importlib-metadata==5.0.0
importlib-resources==5.10.0
incremental==16.10.1
ipython==8.12.2
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.4.1
libclang==14.0.6
localStoragePy==0.2.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
monotonic==1.5
MouseInfo==0.1.3
mss==9.0.1
mysqlclient==1.4.4
neo4j==5.9.0
neo4j-driver==4.4.10
neo4jvis==0.0.1
neobolt==1.7.17
neomodel==5.0.1
netaddr==0.7.19
netifaces==0.10.4
networkx==3.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
paramiko==2.6.0
parso==0.8.3
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
plotly==5.14.1
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.23.2
pure-eval==0.2.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycairo==1.16.2
pycparser==2.19
pycups==1.9.73
pydeck==0.8.0
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.14.1
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.2
pyscreenshot==3.1
PyScreeze==0.1.29
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
semver==2.13.0
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
smmap==5.0.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
sqlalchemy-migrate==0.13.0
sqlparse==0.2.4
st-pages==0.4.1
stack-data==0.6.2
streamlit==1.23.1
streamlit-elements==0.1.0
systemd-python==234
Tempita==0.5.2
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
traitlets==5.9.0
ttystatus==0.38
Twisted==18.9.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.20.0
vboxapi==1.0
wadllib==1.3.3
watchdog==2.1.9
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xkit==0.0.0
zipp==3.10.0
zope.interface==4.7.1
